sigclip/requirements.txt
requests
pillow
huggingface_hub[inference]
python-magic
transformers
torch
diffusers
accelerate
protobuf
sentencepiece
peft
trl
deepspeed
einops
torchvision
opencv-python
usearch